Extract some types from XCScheme.swift
